Override the states library with Drupal 10.2 version in theme
Fix web-241

This puts the states libary override in the themes libraries_override section.
and in the theme.
Has the effect of replacing all instances on the front end with the states
library from Drupal 10.2. Leaves the current 10.3+ version on the admin theme.

This should increase perfomance over loading two state libaries.